requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
drf-yasg>=1.21.0
django-filter>=23.0
scikit-learn>=1.2.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0
python-dotenv>=1.0.0
